Move declarations out of models file in api package
