+++ internal/provider/api/workspace.go
@@ -8,6 +8,12 @@ import (
 	"net/http"
 )
 
+type Workspace struct {
+	ID        string `json:"id"`
+	Name      string `json:"name"`
+	CreatedAt string `json:"created_at"`
+}
+
 type WorkspaceListResponse struct {
 	Workspaces []Workspace `json:"workspaces"`
 }
